+++ packages/terra-core-docs/src/terra-dev-site/doc/form-input/example/uncontrolled/PlaceholderExample.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import Input from 'terra-form-input';
+
+const PlaceholderExample = () => (
+  <Input
+    name="Placeholder"
+    id="uncontrolled-placeholder"
+    placeholder="Enter The Text"
+    showMpageFusionPlaceholder
+  />
+);
+
+export default PlaceholderExample;